BUGFIX: Prevent error on publish workspace via cli
When publishing a workspace via the cli
the redirect package is triggered to
create new redirects if necessary when nodes
have been moved. This causes errors and
prevents the publish because the baseuris
cannot be generated.

This is a hotfix and can be replaced by a
actual working redirect generation.

Relates: #41